% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illuminate_mesh.R
\name{illuminate_mesh}
\alias{illuminate_mesh}
\title{Calculate amount of light hitting triangles in a mesh}
\usage{
illuminate_mesh(mesh, lights, luminance = 1)
}
\arguments{
\item{mesh}{A trimesh object}

\item{lights}{A data.frame with \code{x}, \code{y}, and \code{z} columns giving light
positions in 3D space}

\item{luminance}{The intensity of each light (recycled to the number of rows
in \code{lights})}
}
\value{
A new trimesh object, potentially with additional triangles if
triangles have been splitted. triangle info has been


}
\description{
This functions allows you to calculate the light hitting triangles in a mesh
based on one or more point lights. For each light the visibility of the mesh
from that point is calculated (with potential triangle splitting if it is
only halfway visible) and all visible triangles will get a luminance value
assigned based on it's distance to the light (light fall-off following the
inverse square law). For multiple lights the luminance is cumulative.
}
