% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occlude_mesh.R
\name{occlude_mesh}
\alias{occlude_mesh}
\title{Perform hidden-surface determination of mesh for e.g. occlusion culling}
\usage{
occlude_mesh(mesh, view)
}
\arguments{
\item{mesh}{A trimesh object}

\item{view}{A numeric vector with three elements giving the viewpoint to look
from when calculating occlusion.}
}
\value{
A new trimesh object with a \code{visible} and \code{back_facing} column added
to the triangle info.
}
\description{
This function will determine the visibility of each triangle in a mesh,
potentially cut up triangles if they are only partly occluded. No triangles
are removed by this function but a \code{visible} and \code{back_facing} column is
added indicating if the given triangle can be seen from the specified
viewpoint. If a given triangle is not visible and not back facing it means
that it is being occluded by other triangles in from of it.
}
