% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_mesh.R
\name{project_mesh}
\alias{project_mesh}
\alias{project_coords}
\title{Project the vertices in a mesh onto a plane}
\usage{
project_mesh(mesh, from, to)

project_coords(coords, from, to)
}
\arguments{
\item{mesh}{A trimesh object}

\item{from}{A numeric vector with 3 elements giving the point of view}

\item{to}{A numeric vector with 3 elements giving the point you're looking
towards. The projection plane is defined as a plane including \code{to} and with
\code{to - from} as its normal}

\item{coords}{A data.frame with an \code{x}, \code{y}, and \code{z} column}
}
\value{
Either a new mesh with the vb reprojected, or a new data.frame with
the \code{x}, \code{y}, and \code{z} columns reprojected.
}
\description{
This function cast a ray between a viewpoint and each vertex in a mesh, and
sets the coordinate of the vertex to the intersection of the ray and a plane.
\code{project_coords()} works in the same way but simply takes a data.frame with
\code{x}, \code{y}, and \code{z} values and projcts these
}
