% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimesh_construct.R
\name{trimesh_construct}
\alias{trimesh_construct}
\alias{new_trimesh}
\alias{is_trimesh}
\alias{as_trimesh}
\title{Create a trimesh object}
\usage{
new_trimesh(vertices, triangles, vertex_info = NULL, triangle_info = NULL)

is_trimesh(mesh)

as_trimesh(mesh, ...)
}
\arguments{
\item{vertices}{A 4-row matrix giving the coordinates of vertices in 3D space
(last row is filled with 1's by convention). Each vertex correspond to a
column in the matrix}

\item{triangles}{A 3-row matrix giving the indexes of the three vertices that
corresponds to the triangle. Each triangle corresponds to a column in the
matrix}

\item{vertex_info}{A data.frame giving additional information about each
vertex in the \code{vertices} matrix.}

\item{triangle_info}{A data.frame giving addtional information about each
triangle given in the \code{triangles} matrix}

\item{mesh}{A trimesh or an object convertible to one}

\item{...}{ignored}
}
\description{
These functions helps in creating trimesh objects. A trimesh is a subclass of
mesh3d as defined in the rgl package. It is exclusively using triangles
(unsurprisingly) and adds two additional elements: \code{it_info} which is a
data.frame with information about the triangles in the mesh, and \code{vb_info}
which is a data.frame with information about the vertices. Since trimesh is
all about vector graphics it doesn't concern itself with material, texcoords,
and normals — converting from a mesh3d will loose that information. trimesh
objects can be constructed from either mesh3d object or data frames
containing an \code{x}, \code{y}, and \code{z} column (remaining columns will go to the
\code{it_info} element), or potentially from raw data giving vertice coordinates
and triangle vertex indexes.
}
