% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimesh_index.R
\name{trimesh_index}
\alias{trimesh_index}
\alias{[.trimesh}
\alias{as.data.frame.trimesh}
\alias{triangles}
\alias{vertices}
\alias{triangle_info}
\alias{triangle_info<-}
\alias{vertice_info}
\alias{vertice_info<-}
\alias{mesh_bind}
\title{Extract and set information from trimesh object}
\usage{
\method{[}{trimesh}(x, ..., drop = TRUE)

\method{as.data.frame}{trimesh}(x, row.names = NULL, optional = FALSE, ...)

triangles(mesh)

vertices(mesh)

triangle_info(mesh)

triangle_info(mesh) <- value

vertice_info(mesh)

vertice_info(mesh) <- value

mesh_bind(mesh, ...)
}
\arguments{
\item{x, mesh}{A trimesh object}

\item{...}{additional meshes}

\item{drop}{Should unused vertices be dropped when filtering}

\item{row.names, optional}{ignored}

\item{value}{data.frame with matching number of rows}
}
\value{
Either a new trimesh object, a matrix, or a tibble, depending on the
function
}
\description{
This family of functions allows you to query, set, combine, and filter your
trimesh object.
}
