% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmeshy-package.R
\docType{package}
\name{unmeshy-package}
\alias{unmeshy}
\alias{unmeshy-package}
\title{unmeshy: A Vector Based 3D Renderer}
\description{
Provides utilities for working with objects in 3D and convert these
    to a non-overlaping 2D representation suitable for line-drawing etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/thomasp85/unmeshy}
  \item Report bugs at \url{https://github.com/thomasp85/unmeshy/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

}
\keyword{internal}
